+++ src/js/components/index.js
@@ -21,7 +21,7 @@ export { default as Tree } from './tree.vue';
 export { default as TreeItem } from './tree-item.vue';
 export { default as InputCheck } from './input-check.vue';
 export { default as InputRadio } from './input-radio.vue';
-export { default as Carousel } from 'vue-agile/src/agile.vue';
+export { default as Carousel } from 'vue-agile/src/Agile.vue';
 export { default as Icon } from './icon.vue';
 export { default as CollapseToggle } from './collapse-toggle.vue';
 export { default as Search } from './search.vue';